refactor: schema validation
